Fix extension of the save file dialog on Windows
On Linux, the save file dialog will automatically use the first allowed
extension filter if the current extension doesn't match any of the
selected one.

On Windows, this is not the case and we rename the file name beforhand
with a regex to replace the extension by eml as it is the standard used
also by outlook.

The regex was tested on Linux only by removing the ifdef and ensuring
that the resulting renamedFileName variable is correct.