Fix cleanup of temporary opened files
As owner of the file we need to close our handle before we
can delete the file. All other processes must have opened the
file with at least FILE_SHARE_DELETE so they must live with
the deletion. This also saves the path of the temporary directory
and deletes that, too.

GnuPG-Bug-Id: T6793